requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=1.0.0
xgboost>=1.5.0

# GTFS and transit data
gtfs-realtime-bindings>=0.0.7
protobuf>=3.19.0
requests>=2.26.0

# Reinforcement Learning
gym>=0.21.0
stable-baselines3>=1.6.0
torch>=1.10.0

# Simulation
simpy>=4.0.1
networkx>=2.6.0

# Visualization and Dashboard
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0
dash>=2.0.0
folium>=0.12.0

# Database and Data Processing

sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0

# Utilities
python-dotenv>=0.19.0
tqdm>=4.62.0
geopy>=2.2.0
shapely>=1.8.0